desc
cambiar que si alguien pone el nombre de la foto de otro usuario no se sobrepongan
en el formulario tendria que solucionar bien lo de la busqueda (solo funciona el primer resultado)
en cookie-check hacer un if para que no tenga que buscar en la base de datos cada vez que se accede a un html
agregar una columna que sea quien lo blockeo

+++ PHP/users-update.php
@@ -3,13 +3,14 @@
     include('database.php');
     include('cookie-check.php');
 
-    $target_dir = "/home/dos/public_html/Images/Users/";
+   $target_dir = "/home/dos/public_html/Images/Users/";
 
+    $name = $_FILES["file"]["name"];
     $target_file = $target_dir . basename($_FILES["file"]["name"]);
     $uploadOk = 1;
     $imageFileType = strtolower(pathinfo($target_file,PATHINFO_EXTENSION));
     $target_id = $target_dir.$ids.".".$imageFileType;
-    // Check if image file is a actual image or fake image
+
     if(isset($_POST["submit"])) {
       $check = getimagesize($_FILES["file"]["tmp_name"]);
       if($check !== false) {
@@ -21,42 +22,35 @@
       }
     }
     
-    // Check file size
+
     if ($_FILES["file"]["size"] > 700000) {
       echo "La foto debe pesar menos de 700kb";
       $uploadOk = 0;
     }
-    
-    // Allow certain file formats
+
     if($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg" && $imageFileType != "gif" ) {
       echo "La foto debe ser .jpg .png .jpeg o .gif";
       $uploadOk = 0;
     }
     
-    // Check if $uploadOk is set to 0 by an error
-    if ($uploadOk == 0) {
-    // if everything is ok, try to upload file
-    } else {
-      if (move_uploaded_file($_FILES["file"]["tmp_name"], $target_file)) {
 
-        if($pictures != "default.jpg"){
+    if (!($uploadOk == 0)) {
+      if (move_uploaded_file($_FILES["file"]["tmp_name"], $target_file)) {
+        if($pictures != "default.jpg" && $pictures != $name){
           unlink($target_dir.$pictures);
         }
 
-        rename($target_file,$target_id);
-        $name = $ids.".".$imageFileType;
         $query = "UPDATE users SET picture = '$name' WHERE id='$ids'";
         $result = mysqli_query($connection, $query);
         $_SESSION["picture"] = $name;
-
         if (!$result) {
-          //die('Query Error'. msqli_error($connection));    
+          die('Query Error'. msqli_error($connection));    
         }
-        echo "La foto ". htmlspecialchars( basename( $_FILES["file"]["name"])). " fue subida";
+        echo "La foto de perfil fue subida exitosamente";
       } else {
         echo "Hubo un error subiendo la foto";
       }
     }
-    echo $jsonstring; 
+    echo $jsonstring;
 ?>
